chore(adr-012): clippy clean — remove unnecessary u32 casts in gguf.rs
Two pre-existing `num_layers as u32` casts at GgufBackend::write
(line 288) and write_mmproj_gguf (line 680). `model.metadata.num_layers`
is already declared as u32 in src/ir.rs, so these casts are no-ops
that clippy flags under unnecessary_cast.

Pre-existing issues — not P9b regressions — but this commit clears
the last two warnings in src/backends/gguf.rs since I've been
substantially editing this file in P9b.1. The file is now fully
clippy-clean (zero warnings on -W clippy::all).

Tests still 1118 passing / 0 failing / 9 ignored. No behavior change
(casts on already-typed values are no-ops in machine code).

+++ src/backends/gguf.rs
@@ -285,7 +285,7 @@ impl OutputBackend for GgufBackend {
             let gguf_name = hf_name_to_gguf(
                 name,
                 &resolved_arch_for_tensors,
-                model.metadata.num_layers as u32,
+                model.metadata.num_layers,
             );
             let mut ggml_type = quant_info_to_ggml_type(&qt.quant_info);
 
@@ -677,7 +677,7 @@ fn write_mmproj_gguf(
         let gguf_name = hf_name_to_gguf(
             name,
             &model.metadata.model_type,
-            model.metadata.num_layers as u32,
+            model.metadata.num_layers,
         );
         let ggml_type = quant_info_to_ggml_type(&qt.quant_info);
